requirements.txt
flet>=0.21.0
pyserial>=3.5
opencv-python>=4.8.0
numpy>=1.24.0
requests>=2.31.0 